Expose  InitializeAWSLogging and ShutdownAWSLogging in S3Logging
